# public/openapi.yml
openapi: 3.0.3
info:
  title: 전북바이오 웹 시스템 API
  description: |
    전북특별자치도 바이오산업 육성을 위한 디지털 통합 플랫폼 API
    - 바이오 산업 관련 정보 통합 제공
    - 지원사업, 기술, 인력, 공간, 인증 정보 관리
    - 사용자 권한 기반 접근 제어
  version: 1.0.0
  contact:
    name: 전북테크노파크
    email: support@jeonbuk-bio.kr
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.jeonbuk-bio.kr/v1
    description: Production server
  - url: https://dev-api.jeonbuk-bio.kr/v1
    description: Development server
...
paths:
...
  /users/me/bookmarks/articles:
    get:
      tags: [users]
      summary: 내 기사 북마크 목록 조회
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 북마크한 기사 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Article'
    post:
      tags: [users]
      summary: 기사 북마크 추가
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [articleId]
              properties:
                articleId:
                  type: string
      responses:
        '201':
          description: 북마크 추가 성공
  /users/me/bookmarks/articles/{articleId}:
    delete:
      tags: [users]
      summary: 기사 북마크 삭제
      security:
        - bearerAuth: []
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 북마크 삭제 성공

  # ===== 기업지원사업 API =====
...
  /articles:
    get:
      tags: [articles]
      summary: 기사 목록 조회
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: tag
          in: query
          description: 분야/테마 태그 필터
          schema:
            type: string
        - name: publishDate
          in: query
          description: 발행일 필터 (YYYY-MM)
          schema:
            type: string
            format: date
        - name: sort
          in: query
          description: 정렬 기준
          schema:
            type: string
            enum: [latest, popularity]
            default: latest
      responses:
        '200':
          description: 기사 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Article'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /articles/{id}:
    get:
      tags: [articles]
      summary: 기사 상세 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 기사 상세 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'

  # ===== 성과관리 통계 API =====
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Article:
      type: object
      description: 인터뷰 및 기획기사
      properties:
        id:
          type: string
          example: "article-001"
        title:
          type: string
          example: "전북 바이오텍, AI로 신약 개발 기간 단축"
        author:
          type: string
          example: "이 기자"
        publishDate:
          type: string
          format: date
          example: "2024-08-19"
        tags:
          type: array
          items:
            type: string
          example: ["AI", "신약개발", "스타트업"]
        contentHTML:
          type: string
          example: "<h1>기사 본문</h1><p>내용...</p>"
        images:
          type: array
          items:
            type: string
            format: url
          example: ["https://example.com/image1.jpg"]
        relatedCompanies:
          type: array
          items:
            type: string # Company ID
          example: ["comp-001"]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
...